+++ src/view/button/chip.view.tsx
@@ -26,7 +26,11 @@ export function chipOptions(): ViewData {
             color: 'colors',
             emphasis: {
                 default: 'low',
-                options: ["low", "medium"]
+                options: ["low", "low-normal", "medium"]
+            },
+            size: {
+                default: 'medium',
+                options: ["small", "medium"]
             },
             borderRadius: 'borderRadius',
             disabled: 'boolean',
@@ -53,6 +57,7 @@ export function ChipCreator(props: {p: ChipPropsType}) {
         onChipSelect={(props.p as any).selectable ? () => {} : undefined}
         emphasis={props.p.emphasis}
         borderRadius={props.p.borderRadius}
+        size={props.p.size}
         buttonComponent={
             !(props.p as any).withDropDown
                 ? undefined